+++ packages/components/src/FilterButton/_docs/FilterButtonRemovable.stories.tsx
@@ -1,8 +1,9 @@
 import React from "react"
-import { ComponentMeta, ComponentStory } from "@storybook/react"
+import { Meta, StoryFn } from "@storybook/react"
 import { FilterButtonRemovable } from "../FilterButtonRemovable"
 
 export default {
+  tags: ["autodocs"],
   title: "Components/Filter/Filter Button",
   component: FilterButtonRemovable,
   parameters: {
@@ -13,9 +14,9 @@ export default {
       },
     },
   },
-} as ComponentMeta<typeof FilterButtonRemovable>
+} satisfies Meta<typeof FilterButtonRemovable>
 
-export const FilterButtonRemovableStory: ComponentStory<
+export const FilterButtonRemovableStory: StoryFn<
   typeof FilterButtonRemovable
 > = args => <FilterButtonRemovable {...args} />
 FilterButtonRemovableStory.storyName = "Filter Button Removable"