+++ draft-packages/avatar/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.8.36](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-avatar@2.8.35...@kaizen/draft-avatar@2.8.36) (2023-04-06)
+
+**Note:** Version bump only for package @kaizen/draft-avatar
+
+
+
+
+
 ## [2.8.35](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-avatar@2.8.34...@kaizen/draft-avatar@2.8.35) (2023-04-05)
 
 **Note:** Version bump only for package @kaizen/draft-avatar