+++ draft-packages/select/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.10.34](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-select@2.10.33...@kaizen/draft-select@2.10.34) (2023-04-06)
+
+**Note:** Version bump only for package @kaizen/draft-select
+
+
+
+
+
 ## [2.10.33](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-select@2.10.32...@kaizen/draft-select@2.10.33) (2023-04-05)
 
 **Note:** Version bump only for package @kaizen/draft-select