+++ draft-packages/tile/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.10.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tile@5.10.1...@kaizen/draft-tile@5.10.2) (2023-04-06)
+
+**Note:** Version bump only for package @kaizen/draft-tile
+
+
+
+
+
 ## [5.10.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tile@5.10.0...@kaizen/draft-tile@5.10.1) (2023-04-05)
 
 **Note:** Version bump only for package @kaizen/draft-tile